# docs/swagger.yaml
basePath: /
definitions:
  github_com_dgrco_quikslate_internal_domain.EmployeePosition:
    properties:
      position_id:
        type: string
      user_id:
        type: string
    type: object
  github_com_dgrco_quikslate_internal_domain.LRole:
    enum:
    - employee
    - manager
    - location_lead
    - ""
    - admin
    type: string
    x-enum-comments:
      EmptyRole: no location-scoped role (identity-only sessions)
      LocationLead: manages (leads) all roles at a location
      Manager: manages employees
    x-enum-descriptions:
    - ""
    - manages employees
    - manages (leads) all roles at a location
    - no location-scoped role (identity-only sessions)
    - ""
    x-enum-varnames:
    - Employee
    - Manager
    - LocationLead
    - EmptyRole
    - Admin
...
  github_com_dgrco_quikslate_internal_response.errorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_dgrco_quikslate_internal_service.InviteDTO:
    properties:
      business_name:
        type: string
      email:
        type: string
      expires_at:
        type: string
      role:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.LRole'
    type: object
  internal_handler.AcceptInviteResponse:
    properties:
      business_id:
        type: string
    type: object
...
  internal_handler.CreateInviteResponse:
    properties:
      invite_token:
        type: string
    type: object
...
  internal_handler.MultipleEmployeePositionResponse:
    properties:
      positions:
        items:
          $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.EmployeePosition'
        type: array
    type: object
...
  internal_handler.MultiplePendingInviteResponse:
    properties:
      invites:
        items:
          $ref: '#/definitions/internal_handler.PendingInviteDTO'
        type: array
    type: object
...
  internal_handler.PendingInviteDTO:
    properties:
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      location_id:
        type: string
      role:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.LRole'
    type: object
  internal_handler.SimpleResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_handler.addPositionRequest:
    properties:
      position_id:
        type: string
    type: object
...
  internal_handler.createInviteRequest:
    properties:
      email:
        type: string
      target_role:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.LRole'
    type: object
...
info:
  contact: {}
  description: Scheduling API for businesses, locations, positions, shifts, and employees.
  title: QuikSlate API
  version: "1.0"
paths:
...
  /businesses/{businessId}/employees/{userId}/positions:
    get:
      description: List all positions assigned to a user within the business. Admin,
        Manager, or LocationLead (any location in the business).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultipleEmployeePositionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: List employee positions
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Assign a position to a user. Admin, Manager, or LocationLead (any
        location in the business).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Position to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.addPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Add employee position
      tags:
      - employees
  /businesses/{businessId}/employees/{userId}/positions/{positionId}:
    delete:
      description: Remove a position from a user. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Position ID
        in: path
        name: positionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove employee position
      tags:
      - employees
  /businesses/{businessId}/invites/{token}:
    get:
      description: Preview a pending invite by its raw token (email, business name,
        role, expiry). No authentication required.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Raw invite token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_service.InviteDTO'
        "404":
          description: invite not found, expired, or already accepted
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      summary: Preview invite
      tags:
      - invites
  /businesses/{businessId}/invites/{token}/accept:
    post:
      description: Accept a pending invite as the authenticated caller, joining the
        business at the invited location/role.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Raw invite token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AcceptInviteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: invite not found, expired, or already accepted
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Accept invite
      tags:
      - invites
...
  /businesses/{businessId}/locations/{locationId}/invites:
    get:
      description: List every not-yet-accepted invite at a location. Admin, or LocationLead
        at this location.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultiplePendingInviteResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: List pending invites
      tags:
      - invites
    post:
      consumes:
      - application/json
      description: Invite an email address to join the business at this location with
        a target role. Admin, or LocationLead at this location (may only invite Managers
        or Employees, see canActOnRole).
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Invite details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createInviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CreateInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Create invite
      tags:
      - invites
  /businesses/{businessId}/locations/{locationId}/invites/{inviteId}:
    delete:
      description: Cancel a not-yet-accepted invite at a location, freeing the email
        up to be invited again. Admin, or LocationLead who outranks the invite's target
        role.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Invite ID
        in: path
        name: inviteId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Revoke invite
      tags:
      - invites
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"